+++ src/Symfony/Component/HttpKernel/EventListener/IsSignatureValidAttributeListener.php
@@ -13,7 +13,7 @@ namespace Symfony\Component\HttpKernel\EventListener;
 
 use Psr\Container\ContainerInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-use Symfony\Component\HttpFoundation\UriSigner;
+use Symfony\Component\HttpFoundation\UriSignerInterface;
 use Symfony\Component\HttpKernel\Attribute\IsSignatureValid;
 use Symfony\Component\HttpKernel\Event\ControllerArgumentsEvent;
 use Symfony\Component\HttpKernel\KernelEvents;
@@ -26,7 +26,7 @@ use Symfony\Component\HttpKernel\KernelEvents;
 class IsSignatureValidAttributeListener implements EventSubscriberInterface
 {
     public function __construct(
-        private readonly UriSigner $uriSigner,
+        private readonly UriSignerInterface $uriSigner,
         private readonly ContainerInterface $container,
     ) {
     }
@@ -50,8 +50,8 @@ class IsSignatureValidAttributeListener implements EventSubscriberInterface
             }
 
             $signer = $this->container->get($attribute->signer);
-            if (!$signer instanceof UriSigner) {
-                throw new \LogicException(\sprintf('The service "%s" is not an instance of "%s".', $attribute->signer, UriSigner::class));
+            if (!$signer instanceof UriSignerInterface) {
+                throw new \LogicException(\sprintf('The service "%s" is not an instance of "%s".', $attribute->signer, UriSignerInterface::class));
             }
 
             $signer->verify($request);